refactor: rename components for uniformity
